clkmgr: Improve markdown documents
Replace the sign characters in HLD_clkmgr.md,
 so we use only proper markdown tables characters.

markdown follow html, a new line in file,
 do not create line feed in rendered document.
To have line feed, we need to add 2 spaces, or an empty line.

Fix the link to sample JSON configuration file.

Index are increase automatically, we just use '1.'
 and the rendering will increase the numbers for us.
So in case we insert a new row (using '1.'),
 the numbers will increase accordingly!

Signed-off-by: Erez Geva <ErezGeva2@gmail.com>